Crate google_blogger3

This documentation was generated from blogger crate version 1.0.9+20150422, where 20150422 is the exact revision of the blogger:v3 schema built by the mako code generator v1.0.9.

Everything else about the blogger v3 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

blog user infos

get

blogs

get, get by url and list by user

comments

approve, delete, get, list, list by blog, mark as spam and remove content

page views

get

pages

delete, get, insert, list, patch, publish, revert and update

post user infos

get and list

posts

delete, get, get by path, insert, list, patch, publish, revert, search and update

users

get

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.posts().insert(...).doit()
let r = hub.posts().publish(...).doit()
let r = hub.posts().delete(...).doit()
let r = hub.posts().revert(...).doit()
let r = hub.posts().update(...).doit()
let r = hub.posts().get_by_path(...).doit()
let r = hub.posts().search(...).doit()
let r = hub.posts().patch(...).doit()
let r = hub.posts().list(...).doit()
let r = hub.posts().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-blogger3 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_blogger3 as blogger3;
use blogger3::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use blogger3::Blogger;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Blogger::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.posts().list("blogId")
             .view("labore")
             .add_status("sea")
             .start_date("nonumy")
             .page_token("dolores")
             .order_by("gubergren")
             .max_results(6)
             .labels("aliquyam")
             .fetch_images(false)
             .fetch_bodies(false)
             .end_date("justo")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Blog
There is no detailed description.

BlogGetByUrlCall
Retrieve a Blog by URL.

BlogGetCall
Gets one blog by ID.

BlogList
There is no detailed description.

BlogListByUserCall
Retrieves a list of blogs, possibly filtered.

BlogLocale
The locale this Blog is set to.

BlogMethods
A builder providing access to all methods supported on blog resources. It is not used directly, but through the Blogger hub.

BlogPages
The container of pages in this blog.

BlogPerUserInfo
There is no detailed description.

BlogPosts
The container of posts in this blog.

BlogUserInfo
There is no detailed description.

BlogUserInfoGetCall
Gets one blog and user info pair by blogId and userId.

BlogUserInfoMethods
A builder providing access to all methods supported on blogUserInfo resources. It is not used directly, but through the Blogger hub.

Blogger
Central instance to access all Blogger related resource activities

Chunk

Comment
There is no detailed description.

CommentApproveCall
Marks a comment as not spam.

CommentAuthor
The author of this Comment.

CommentAuthorImage
The comment creator's avatar.

CommentBlog
Data about the blog containing this comment.

CommentDeleteCall
Delete a comment by ID.

CommentGetCall
Gets one comment by ID.

CommentInReplyTo
Data about the comment this is in reply to.

CommentList
There is no detailed description.

CommentListByBlogCall
Retrieves the comments for a blog, across all posts, possibly filtered.

CommentListCall
Retrieves the comments for a post, possibly filtered.

CommentMarkAsSpamCall
Marks a comment as spam.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the Blogger hub.

CommentPost
Data about the post containing this comment.

CommentRemoveContentCall
Removes the content of a comment.

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DummyNetworkStream

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

JsonServerError
A utility type which can decode a server response that indicates error

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Page
There is no detailed description.

PageAuthor
The author of this Page.

PageAuthorImage
The page author's avatar.

PageBlog
Data about the blog containing this Page.

PageDeleteCall
Delete a page by ID.

PageGetCall
Gets one blog page by ID.

PageInsertCall
Add a page.

PageList
There is no detailed description.

PageListCall
Retrieves the pages for a blog, optionally including non-LIVE statuses.

PageMethods
A builder providing access to all methods supported on page resources. It is not used directly, but through the Blogger hub.

PagePatchCall
Update a page. This method supports patch semantics.

PagePublishCall
Publishes a draft page.

PageRevertCall
Revert a published or scheduled page to draft state.

PageUpdateCall
Update a page.

PageViewGetCall
Retrieve pageview stats for a Blog.

PageViewMethods
A builder providing access to all methods supported on pageView resources. It is not used directly, but through the Blogger hub.

Pageviews
There is no detailed description.

PageviewsCounts
The container of posts in this blog.

Post
There is no detailed description.

PostAuthor
The author of this Post.

PostAuthorImage
The Post author's avatar.

PostBlog
Data about the blog containing this Post.

PostDeleteCall
Delete a post by ID.

PostGetByPathCall
Retrieve a Post by Path.

PostGetCall
Get a post by ID.

PostImages
Display image for the Post.

PostInsertCall
Add a post.

PostList
There is no detailed description.

PostListCall
Retrieves a list of posts, possibly filtered.

PostLocation
The location for geotagged posts.

PostMethods
A builder providing access to all methods supported on post resources. It is not used directly, but through the Blogger hub.

PostPatchCall
Update a post. This method supports patch semantics.

PostPerUserInfo
There is no detailed description.

PostPublishCall
Publishes a draft post, optionally at the specific time of the given publishDate parameter.

PostReplies
The container of comments on this Post.

PostRevertCall
Revert a published or scheduled post to draft state.

PostSearchCall
Search for a post.

PostUpdateCall
Update a post.

PostUserInfo
There is no detailed description.

PostUserInfoGetCall
Gets one post and user info pair, by post ID and user ID. The post user info contains per-user information about the post, such as access rights, specific to the user.

PostUserInfoListCall
Retrieves a list of post and post user info pairs, possibly filtered. The post user info contains per-user information about the post, such as access rights, specific to the user.

PostUserInfoMethods
A builder providing access to all methods supported on postUserInfo resources. It is not used directly, but through the Blogger hub.

PostUserInfosList
There is no detailed description.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

User
There is no detailed description.

UserBlogs
The container of blogs for this user.

UserGetCall
Gets one user by ID.

UserLocale
This user's locale

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the Blogger hub.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.