Module google_blogger3::api

Structs

Blog
There is no detailed description.

BlogGetByUrlCall
Gets a blog by url.

BlogGetCall
Gets a blog by id.

BlogList
There is no detailed description.

BlogListByUserCall
Lists blogs by user.

BlogLocale
The locale this Blog is set to.

BlogMethods
A builder providing access to all methods supported on blog resources. It is not used directly, but through the Blogger hub.

BlogPages
The container of pages in this blog.

BlogPerUserInfo
There is no detailed description.

BlogPosts
The container of posts in this blog.

BlogUserInfo
There is no detailed description.

BlogUserInfoGetCall
Gets one blog and user info pair by blog id and user id.

BlogUserInfoMethods
A builder providing access to all methods supported on blogUserInfo resources. It is not used directly, but through the Blogger hub.

Blogger
Central instance to access all Blogger related resource activities

Comment
There is no detailed description.

CommentApproveCall
Marks a comment as not spam by blog id, post id and comment id.

CommentAuthor
The author of this Comment.

CommentAuthorImage
The creator's avatar.

CommentBlog
Data about the blog containing this comment.

CommentDeleteCall
Deletes a comment by blog id, post id and comment id.

CommentGetCall
Gets a comment by id.

CommentInReplyTo
Data about the comment this is in reply to.

CommentList
There is no detailed description.

CommentListByBlogCall
Lists comments by blog.

CommentListCall
Lists comments.

CommentMarkAsSpamCall
Marks a comment as spam by blog id, post id and comment id.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the Blogger hub.

CommentPost
Data about the post containing this comment.

CommentRemoveContentCall
Removes the content of a comment by blog id, post id and comment id.

Page
There is no detailed description.

PageAuthor
The author of this Page.

PageAuthorImage
The creator's avatar.

PageBlog
Data about the blog containing this Page.

PageDeleteCall
Deletes a page by blog id and page id.

PageGetCall
Gets a page by blog id and page id.

PageInsertCall
Inserts a page.

PageList
There is no detailed description.

PageListCall
Lists pages.

PageMethods
A builder providing access to all methods supported on page resources. It is not used directly, but through the Blogger hub.

PagePatchCall
Patches a page.

PagePublishCall
Publishes a page.

PageRevertCall
Reverts a published or scheduled page to draft state.

PageUpdateCall
Updates a page by blog id and page id.

PageViewGetCall
Gets page views by blog id.

PageViewMethods
A builder providing access to all methods supported on pageView resources. It is not used directly, but through the Blogger hub.

Pageviews
There is no detailed description.

PageviewsCounts
The container of posts in this blog.

Post
There is no detailed description.

PostAuthor
The author of this Post.

PostAuthorImage
The creator's avatar.

PostBlog
Data about the blog containing this Post.

PostDeleteCall
Deletes a post by blog id and post id.

PostGetByPathCall
Gets a post by path.

PostGetCall
Gets a post by blog id and post id

PostImages
Display image for the Post.

PostInsertCall
Inserts a post.

PostList
There is no detailed description.

PostListCall
Lists posts.

PostLocation
The location for geotagged posts.

PostMethods
A builder providing access to all methods supported on post resources. It is not used directly, but through the Blogger hub.

PostPatchCall
Patches a post.

PostPerUserInfo
There is no detailed description.

PostPublishCall
Publishes a post.

PostReplies
The container of comments on this Post.

PostRevertCall
Reverts a published or scheduled post to draft state.

PostSearchCall
Searches for posts matching given query terms in the specified blog.

PostUpdateCall
Updates a post by blog id and post id.

PostUserInfo
There is no detailed description.

PostUserInfoGetCall
Gets one post and user info pair, by post_id and user_id.

PostUserInfoListCall
Lists post and user info pairs.

PostUserInfoMethods
A builder providing access to all methods supported on postUserInfo resources. It is not used directly, but through the Blogger hub.

PostUserInfosList
There is no detailed description.

User
There is no detailed description.

UserBlogs
The container of blogs for this user.

UserGetCall
Gets one user by user_id.

UserLocale
This user's locale

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the Blogger hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.