Struct google_blogger3::Blog

pub struct Blog {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub description: Option<String>,
    pub locale: Option<BlogLocale>,
    pub custom_meta_data: Option<String>,
    pub posts: Option<BlogPosts>,
    pub updated: Option<String>,
    pub id: Option<String>,
    pub url: Option<String>,
    pub published: Option<String>,
    pub pages: Option<BlogPages>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list by user blogs (none)

get blogs (response)

get by url blogs (response)

Fields

status

The status of the blog.

kind

The kind of this entry. Always blogger#blog

description

The description of this blog. This is displayed underneath the title.

locale

The locale this Blog is set to.

custom_meta_data

The JSON custom meta-data for the Blog

posts

The container of posts in this blog.

updated

RFC 3339 date-time when this blog was last updated.

id

The identifier for this resource.

url

The URL where this blog is published.

published

RFC 3339 date-time when this blog was published.

pages

The container of pages in this blog.

self_link

The API REST URL to fetch this resource from.

name

The name of this blog. This is displayed as the title.