Struct google_blogger3::BlogMethods

pub struct BlogMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on blog resources. It is not used directly, but through the Blogger hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_blogger3 as blogger3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use blogger3::Blogger;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Blogger::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `get_by_url(...)` and `list_by_user(...)`
// to build up your call.
let rb = hub.blogs();

Methods

impl<'a, C, A> BlogMethods<'a, C, A>

fn list_by_user(&self, user_id: &str) -> BlogListByUserCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of blogs, possibly filtered.

Arguments

userId - ID of the user whose blogs are to be fetched. Either the word 'self' (sans quote marks) or the user's profile identifier.

fn get(&self, blog_id: &str) -> BlogGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets one blog by ID.

Arguments

blogId - The ID of the blog to get.

fn get_by_url(&self, url: &str) -> BlogGetByUrlCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieve a Blog by URL.

Arguments

url - The URL of the blog to retrieve.