Struct google_blogger3::CommentMethods

pub struct CommentMethods<'a, C, A> where C: 'a, A: 'a { /* fields omitted */ }

A builder providing access to all methods supported on comment resources. It is not used directly, but through the Blogger hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_blogger3 as blogger3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use blogger3::Blogger;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Blogger::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `approve(...)`, `delete(...)`, `get(...)`, `list(...)`, `list_by_blog(...)`, `mark_as_spam(...)` and `remove_content(...)`
// to build up your call.
let rb = hub.comments();

Methods

impl<'a, C, A> CommentMethods<'a, C, A>

fn get(&self,
blog_id: &str,
post_id: &str,
comment_id: &str)
-> CommentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets one comment by ID.

Arguments

blogId - ID of the blog to containing the comment.

postId - ID of the post to fetch posts from.

commentId - The ID of the comment to get.

fn remove_content(&self,
blog_id: &str,
post_id: &str,
comment_id: &str)
-> CommentRemoveContentCall<'a, C, A>

Create a builder to help you perform the following task:

Removes the content of a comment.

Arguments

blogId - The ID of the Blog.

postId - The ID of the Post.

commentId - The ID of the comment to delete content from.

fn list_by_blog(&self, blog_id: &str) -> CommentListByBlogCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the comments for a blog, across all posts, possibly filtered.

Arguments

blogId - ID of the blog to fetch comments from.

fn mark_as_spam(&self,
blog_id: &str,
post_id: &str,
comment_id: &str)
-> CommentMarkAsSpamCall<'a, C, A>

Create a builder to help you perform the following task:

Marks a comment as spam.

Arguments

blogId - The ID of the Blog.

postId - The ID of the Post.

commentId - The ID of the comment to mark as spam.

fn list(&self, blog_id: &str, post_id: &str) -> CommentListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the comments for a post, possibly filtered.

Arguments

blogId - ID of the blog to fetch comments from.

postId - ID of the post to fetch posts from.

fn approve(&self,
blog_id: &str,
post_id: &str,
comment_id: &str)
-> CommentApproveCall<'a, C, A>

Create a builder to help you perform the following task:

Marks a comment as not spam.

Arguments

blogId - The ID of the Blog.

postId - The ID of the Post.

commentId - The ID of the comment to mark as not spam.

fn delete(&self,
blog_id: &str,
post_id: &str,
comment_id: &str)
-> CommentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a comment by ID.

Arguments

blogId - The ID of the Blog.

postId - The ID of the Post.

commentId - The ID of the comment to delete.