Struct google_blogger3::Page

pub struct Page {
    pub status: Option<String>,
    pub blog: Option<PageBlog>,
    pub kind: Option<String>,
    pub title: Option<String>,
    pub url: Option<String>,
    pub author: Option<PageAuthor>,
    pub updated: Option<String>,
    pub content: Option<String>,
    pub etag: Option<String>,
    pub published: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update pages (request|response)

get pages (response)

insert pages (request|response)

publish pages (response)

list pages (none)

patch pages (request|response)

revert pages (response)

delete pages (none)

Fields

status

The status of the page for admin resources (either LIVE or DRAFT).

blog

Data about the blog containing this Page.

kind

The kind of this entity. Always blogger#page

title

The title of this entity. This is the name displayed in the Admin user interface.

url

The URL that this Page is displayed at.

author

The author of this Page.

updated

RFC 3339 date-time when this Page was last updated.

content

The body content of this Page, in HTML.

etag

Etag of the resource.

published

RFC 3339 date-time when this Page was published.

id

The identifier for this resource.

self_link

The API REST URL to fetch this resource from.