Struct google_blogger3::PageViewMethods

pub struct PageViewMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on pageView resources. It is not used directly, but through the Blogger hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_blogger3 as blogger3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use blogger3::Blogger;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Blogger::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`
// to build up your call.
let rb = hub.page_views();

Methods

impl<'a, C, A> PageViewMethods<'a, C, A>

fn get(&self, blog_id: &str) -> PageViewGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieve pageview stats for a Blog.

Arguments

blogId - The ID of the blog to get.