Struct google_blogger3::PostMethods

pub struct PostMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on post resources. It is not used directly, but through the Blogger hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_blogger3 as blogger3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use blogger3::Blogger;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Blogger::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `get_by_path(...)`, `insert(...)`, `list(...)`, `patch(...)`, `publish(...)`, `revert(...)`, `search(...)` and `update(...)`
// to build up your call.
let rb = hub.posts();

Methods

impl<'a, C, A> PostMethods<'a, C, A>

fn update(&self, request: Post, blog_id: &str, post_id: &str) -> PostUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update a post.

Arguments

request - No description provided.

blogId - The ID of the Blog.

postId - The ID of the Post.

fn get_by_path(&self, blog_id: &str, path: &str) -> PostGetByPathCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieve a Post by Path.

Arguments

blogId - ID of the blog to fetch the post from.

path - Path of the Post to retrieve.

fn get(&self, blog_id: &str, post_id: &str) -> PostGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get a post by ID.

Arguments

blogId - ID of the blog to fetch the post from.

postId - The ID of the post

fn insert(&self, request: Post, blog_id: &str) -> PostInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Add a post.

Arguments

request - No description provided.

blogId - ID of the blog to add the post to.

fn publish(&self, blog_id: &str, post_id: &str) -> PostPublishCall<'a, C, A>

Create a builder to help you perform the following task:

Publishes a draft post, optionally at the specific time of the given publishDate parameter.

Arguments

blogId - The ID of the Blog.

postId - The ID of the Post.

fn delete(&self, blog_id: &str, post_id: &str) -> PostDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a post by ID.

Arguments

blogId - The ID of the Blog.

postId - The ID of the Post.

fn revert(&self, blog_id: &str, post_id: &str) -> PostRevertCall<'a, C, A>

Create a builder to help you perform the following task:

Revert a published or scheduled post to draft state.

Arguments

blogId - The ID of the Blog.

postId - The ID of the Post.

fn search(&self, blog_id: &str, q: &str) -> PostSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Search for a post.

Arguments

blogId - ID of the blog to fetch the post from.

q - Query terms to search this blog for matching posts.

fn patch(&self, request: Post, blog_id: &str, post_id: &str) -> PostPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update a post. This method supports patch semantics.

Arguments

request - No description provided.

blogId - The ID of the Blog.

postId - The ID of the Post.

fn list(&self, blog_id: &str) -> PostListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of posts, possibly filtered.

Arguments

blogId - ID of the blog to fetch posts from.