Struct google_blogger3::PostUpdateCall

pub struct PostUpdateCall<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

Update a post.

A builder for the update method supported by a post resource. It is not used directly, but through a PostMethods instance.

Example

Instantiate a resource method builder

use blogger3::Post;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Post::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.posts().update(req, "blogId", "postId")
             .revert(true)
             .publish(true)
             .max_comments(31)
             .fetch_images(true)
             .fetch_body(true)
             .doit();

Methods

impl<'a, C, A> PostUpdateCall<'a, C, A> where
C: BorrowMut<Client>,
A: GetToken,

pub fn doit(self) -> Result<(Response, Post)>

Perform the operation you have build so far.

pub fn request(self, new_value: Post) -> PostUpdateCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn blog_id(self, new_value: &str) -> PostUpdateCall<'a, C, A>

The ID of the Blog.

Sets the blog id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn post_id(self, new_value: &str) -> PostUpdateCall<'a, C, A>

The ID of the Post.

Sets the post id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn revert(self, new_value: bool) -> PostUpdateCall<'a, C, A>

Whether a revert action should be performed when the post is updated (default: false).

Sets the revert query property to the given value.

pub fn publish(self, new_value: bool) -> PostUpdateCall<'a, C, A>

Whether a publish action should be performed when the post is updated (default: false).

Sets the publish query property to the given value.

pub fn max_comments(self, new_value: u32) -> PostUpdateCall<'a, C, A>

Maximum number of comments to retrieve with the returned post.

Sets the max comments query property to the given value.

pub fn fetch_images(self, new_value: bool) -> PostUpdateCall<'a, C, A>

Whether image URL metadata for each post is included in the returned result (default: false).

Sets the fetch images query property to the given value.

pub fn fetch_body(self, new_value: bool) -> PostUpdateCall<'a, C, A>

Whether the body content of the post is included with the result (default: true).

Sets the fetch body query property to the given value.

pub fn delegate(
self,
new_value: &'a mut dyn Delegate
) -> PostUpdateCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> PostUpdateCall<'a, C, A> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - Deprecated. Please use quotaUser instead.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

pub fn add_scope<T, S>(self, scope: T) -> PostUpdateCall<'a, C, A> where
T: Into<Option<S>>,
S: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope. If None is specified, then all scopes will be removed and no default scope will be used either. In that case, you have to specify your API-key using the key parameter (see the param() function for details).

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.