Struct google_blogger3::Page

pub struct Page {
    pub status: Option<String>,
    pub blog: Option<PageBlog>,
    pub kind: Option<String>,
    pub title: Option<String>,
    pub url: Option<String>,
    pub author: Option<PageAuthor>,
    pub updated: Option<String>,
    pub content: Option<String>,
    pub etag: Option<String>,
    pub published: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update pages (request|response)

get pages (response)

insert pages (request|response)

publish pages (response)

list pages (none)

patch pages (request|response)

revert pages (response)

delete pages (none)

Fields
status: Option<String>

The status of the page for admin resources (either LIVE or DRAFT).

blog: Option<PageBlog>

Data about the blog containing this Page.

kind: Option<String>

The kind of this entity. Always blogger#page

title: Option<String>

The title of this entity. This is the name displayed in the Admin user interface.

url: Option<String>

The URL that this Page is displayed at.

author: Option<PageAuthor>

The author of this Page.

updated: Option<String>

RFC 3339 date-time when this Page was last updated.

content: Option<String>

The body content of this Page, in HTML.

etag: Option<String>

Etag of the resource.

published: Option<String>

RFC 3339 date-time when this Page was published.

id: Option<String>

The identifier for this resource.

self_link: Option<String>

The API REST URL to fetch this resource from.