Struct google_blogger3::Post

pub struct Post {
    pub status: Option<String>,
    pub updated: Option<String>,
    pub reader_comments: Option<String>,
    pub labels: Option<Vec<String>>,
    pub replies: Option<PostReplies>,
    pub images: Option<Vec<PostImages>>,
    pub id: Option<String>,
    pub blog: Option<PostBlog>,
    pub kind: Option<String>,
    pub title_link: Option<String>,
    pub author: Option<PostAuthor>,
    pub url: Option<String>,
    pub title: Option<String>,
    pub custom_meta_data: Option<String>,
    pub content: Option<String>,
    pub etag: Option<String>,
    pub location: Option<PostLocation>,
    pub published: Option<String>,
    pub self_link: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert posts (request|response)

publish posts (response)

delete posts (none)

revert posts (response)

update posts (request|response)

get by path posts (response)

search posts (none)

patch posts (request|response)

list posts (none)

get posts (response)

Fields
status: Option<String>

Status of the post. Only set for admin-level requests

updated: Option<String>

RFC 3339 date-time when this Post was last updated.

reader_comments: Option<String>

Comment control and display setting for readers of this post.

labels: Option<Vec<String>>

The list of labels this Post was tagged with.

replies: Option<PostReplies>

The container of comments on this Post.

images: Option<Vec<PostImages>>

Display image for the Post.

id: Option<String>

The identifier of this Post.

blog: Option<PostBlog>

Data about the blog containing this Post.

kind: Option<String>

The kind of this entity. Always blogger#post

title_link: Option<String>

The title link URL, similar to atom's related link.

author: Option<PostAuthor>

The author of this Post.

url: Option<String>

The URL where this Post is displayed.

title: Option<String>

The title of the Post.

custom_meta_data: Option<String>

The JSON meta-data for the Post.

content: Option<String>

The content of the Post. May contain HTML markup.

etag: Option<String>

Etag of the resource.

location: Option<PostLocation>

The location for geotagged posts.

published: Option<String>

RFC 3339 date-time when this Post was published.

self_link: Option<String>

The API REST URL to fetch this resource from.